refactor: dependecy inversion for testability
